"Low-level" materials added. Also added dynamic resources.
D3D12_ROOT_SIGNATURE_FLAG_SAMPLER_HEAP_DIRECTLY_INDEXED requires to add sampler_desc_heap in Core module since that's where we set descriptor heaps per pass/draw call. (though it has no use whatsoever now). Might be a subject  to change in a future (and maybe some other submodule (e.g. gpass) will set it, but I doubt that).

+++ Engine/Graphics/Direct3D12/D3D12CommonHeaders.h
@@ -51,7 +51,7 @@ if(FAILED(x)) {                                 \
 #define NAME_D3D12_OBJECT_INDEXED(obj, index, name)                 \
 {                                                                   \
 wchar_t full_name[128];                                             \
-if(swprintf_s(full_name, L"%s[%u]", name, index) > 0) {             \
+if(swprintf_s(full_name, L"%s[%llu]", name, (u64)index) > 0) {      \
     obj->SetName(full_name);                                        \
     OutputDebugString(L"::D3D12 Object Created: ");                 \
     OutputDebugString(full_name);                                   \
